Remove foreign renderer and onscreen setters
Now that we have explicit constructors, we can remove the setters.
At the same time, we add a method on CoglDisplay, to query a
suitable XVisualInfo for creating X windows. Also, since we no
longer pass width and height to the onscreen constructors, we need
to update the viewport when the window size changes.