Stop creating native windows
Make it painfully clear that all windows must be provided by the
integrating toolkit, and drop all code that dealt with creating
windows. Onscreen constructors are no longer "foreign", and the
fields in CoglOnscreen are the only place where XIDs, HWNDs and
wl_surfaces are stored.
The main reason for this is that soon Cogl will stop dealing with
all events (except maybe GLX swap events), requiring instead the
integrating toolkit to update the framebuffer size and to report
damage to CoglTextureX11Pixmap.
As a side effect, this solves the problem of duplicate events in
the win32 winsys.