Replace hello world example with GDK example
Since we can no longer create a CoglOnscreen with cogl_onscreen_new,
we need a toolkit to create the window to render. This shows how
to use Cogl to render in GDK window (until GDK learns to do this
itself, like it did with cairo).
The crate example instead was blocked, and ported to the native
EGL platform.